Move iw scraping for local wifi config into a shared utility
